Add info logging for responses
