+++ packages/core/src/base/staticClient.ts
@@ -80,13 +80,23 @@ export function getHeader(
 }
 
 /**
- * Log the request being made.
- * 
+ * Format the request being made for logging.
+ *
  * @param resource The resource being requested
  * @param fetchOptions The options to the fetch call
  */
 export const formatFetchForInfoLog = 
   (resource:string, fetchOptions:RequestInit): string => `Request: ${fetchOptions.method.toUpperCase()} ${resource}`;
+
+/**
+ * Format the response received for logging.
+ *
+ * @param response The response received
+ */
+export const formatResponseForInfoLog = (response:Response): string => {
+  const successString = (response.ok || response.status === 304) ? "successful" : "unsuccessful";
+  return `Received ${successString} response: ${response.status} ${response.statusText}`;
+};
 
 /**
  * Makes an HTTP call specified by the method parameter with the options passed.
@@ -152,6 +162,8 @@ async function runFetch(
   sdkLogger.info(formatFetchForInfoLog(resource, fetchOptions));
 
   const response = await fetch(resource, fetchOptions);
+
+  sdkLogger.info(formatResponseForInfoLog(response));
 
   return options.rawResponse ? response : getObjectFromResponse(response);
 }
